Updated the name of algorithm
